+++ src/xulbux/regex.py
@@ -1,10 +1,6 @@
-from typing import TypeAlias, Optional
+from typing import Optional
 import regex as _rx
 import re as _re
-
-
-Pattern: TypeAlias = _re.Pattern[str] | _rx.Pattern[str]
-Match: TypeAlias = _re.Match[str] | _rx.Match[str]
 
 
 class Regex: